refactor: 멀티모듈 적용. * 프로젝트 분리
